+++ back-end/prescrypt/src/main/java/org/H4212/services/ServiceUser.java
@@ -72,7 +72,7 @@ public class ServiceUser {
 
         if(resultSetDoctor.next())
         {
-            return new Doctor((long) resultSetDoctor.getInt(1), resultSetDoctor.getString(2), resultSetDoctor.getString(3), (long) resultSetDoctor.getInt(4), resultSetDoctor.getString(5), resultSetDoctor.getString(6), resultSetDoctor.getString(7), resultSetDoctor.getString(8));
+            return new Doctor((long) resultSetDoctor.getLong(1), resultSetDoctor.getString(2), resultSetDoctor.getString(3), (long) resultSetDoctor.getLong(4), resultSetDoctor.getString(5), resultSetDoctor.getString(6), resultSetDoctor.getString(7), resultSetDoctor.getString(8));
         }else{
             System.out.println("ResultSet is empty");
             return new Doctor();
@@ -119,7 +119,7 @@ public class ServiceUser {
 
         if(resultSetDoctor.next())
         {
-            return new Doctor((long) resultSetDoctor.getInt(1), resultSetDoctor.getString(2), resultSetDoctor.getString(3), (long) resultSetDoctor.getInt(4), resultSetDoctor.getString(5), resultSetDoctor.getString(6), resultSetDoctor.getString(7), resultSetDoctor.getString(8));
+            return new Doctor((long) resultSetDoctor.getLong(1), resultSetDoctor.getString(2), resultSetDoctor.getString(3), (long) resultSetDoctor.getLong(4), resultSetDoctor.getString(5), resultSetDoctor.getString(6), resultSetDoctor.getString(7), resultSetDoctor.getString(8));
         }else{
             System.out.println("ResultSetDoctor is empty");
             return new Doctor();
@@ -170,7 +170,7 @@ public class ServiceUser {
 
         if(resultSetPharmacist.next())
         {
-            return new Pharmacist((long) resultSetPharmacist.getInt(1), resultSetPharmacist.getString(2), resultSetPharmacist.getString(3));
+            return new Pharmacist((long) resultSetPharmacist.getLong(1), resultSetPharmacist.getString(2), resultSetPharmacist.getString(3));
         }else{
             System.out.println("ResultSet is empty");
             return new Pharmacist();
@@ -198,7 +198,7 @@ public class ServiceUser {
 
         if(resultSet.next())
         {
-            return new Person((long) resultSet.getInt(1), resultSet.getString(2), resultSet.getString(2));
+            return new Person((long) resultSet.getLong(1), resultSet.getString(2), resultSet.getString(2));
         }else{
             System.out.println("ResultSet is empty");
             return new Person();
@@ -222,7 +222,7 @@ public class ServiceUser {
 
         if(resultSet.next())
         {
-            return new Patient((long) resultSet.getInt(1), resultSet.getString(2), resultSet.getString(3), resultSet.getInt(4), resultSet.getInt(5), resultSet.getBoolean(6), resultSet.getInt(7));
+            return new Patient((long) resultSet.getLong(1), resultSet.getString(2), resultSet.getString(3), resultSet.getInt(4), resultSet.getInt(5), resultSet.getBoolean(6), resultSet.getInt(7));
         }else{
             System.out.println("ResultSet is empty");
             return new Patient();
@@ -294,7 +294,7 @@ public class ServiceUser {
         preparedStatementPharmacist.executeUpdate();
     }
 
-    public void deleteDoctor(Long doctorId) throws SQLException{
+    public void deleteDoctor(long doctorId) throws SQLException{
 
         String stringQueryDoctor =
                 """
@@ -318,7 +318,7 @@ public class ServiceUser {
 
     }
 
-    public void deletePharmacist(Long pharmacistId) throws SQLException{
+    public void deletePharmacist(long pharmacistId) throws SQLException{
 
         String stringQueryPharmacist =
                 """
